+++ source/Refactorings/Refactorings/IfStatementRefactoring.cs
@@ -37,8 +37,8 @@ namespace Roslynator.CSharp.Refactorings
                 if (context.IsRefactoringEnabled(RefactoringIdentifiers.SwapStatementsInIfElse))
                     SwapStatementInIfElseRefactoring.ComputeRefactoring(context, ifStatement);
 
-                if (context.IsRefactoringEnabled(RefactoringIdentifiers.ReplaceIfElseWithSwitch))
-                    await ReplaceIfElseWithSwitchRefactoring.ComputeRefactoringAsync(context, ifStatement).ConfigureAwait(false);
+                if (context.IsRefactoringEnabled(RefactoringIdentifiers.ReplaceIfWithSwitch))
+                    await ReplaceIfWithSwitchRefactoring.ComputeRefactoringAsync(context, ifStatement).ConfigureAwait(false);
 
                 if (context.IsRefactoringEnabled(RefactoringIdentifiers.SplitIfStatement))
                     SplitIfStatementRefactoring.ComputeRefactoring(context, ifStatement);
@@ -58,8 +58,21 @@ namespace Roslynator.CSharp.Refactorings
                 {
                     context.RegisterRefactoring(
                         "Reduce if nesting",
-                        cancellationToken => ReduceIfNestingRefactoring.RefactorAsync(context.Document, ifStatement, context.CancellationToken));
+                        cancellationToken => ReduceIfNestingRefactoring.RefactorAsync(context.Document, ifStatement, false, context.CancellationToken));
+
+                    if (ReduceIfNestingRefactoring.IsFixableRecursively(ifStatement))
+                    {
+                        context.RegisterRefactoring(
+                            "Reduce if nesting (recursively)",
+                            cancellationToken => ReduceIfNestingRefactoring.RefactorAsync(context.Document, ifStatement, true, context.CancellationToken));
+                    }
                 }
+            }
+
+            if (context.IsRefactoringEnabled(RefactoringIdentifiers.ReplaceIfElseWithIfReturn)
+                && context.Span.IsEmptyAndContainedInSpanOrBetweenSpans(ifStatement.IfKeyword))
+            {
+                ReplaceIfElseWithIfReturnRefactoring.ComputeRefactoring(context, ifStatement);
             }
         }
     }