+++ ChangeLog.md
@@ -1,3 +1,27 @@
+### 1.4.58 (2017-09-16)
+
+#### Analyzers
+
+  * Remove analyzer UseCSharp6DictionaryInitializer (RCS1095)
+
+#### Refactorings
+
+##### New Refactorings
+
+  * UseCSharp6DictionaryInitializer (RR0191)
+
+### 1.4.57 (2017-09-06)
+
+#### Refactorings
+
+##### New Refactorings
+
+  * ReplaceIfElseWithIfReturn (RR0190)
+
+#### Code Fixes
+
+  * Add code fix for CS0021.
+
 ### 1.4.56 (2017-08-28)
 
 #### Analyzers